+++ scripts/ib_data_reset.py
@@ -17,68 +17,13 @@
 '''
 IB api client data feed reset hack for i3.
 
+Delegates to ``i3ipc`` python lib to detect the correct local
+window to click-activate and ``xdotool`` to send the mouse
+events to said window.
+
 '''
-import subprocess
+from piker.brokers.ib._util import i3ipc_xdotool_manual_click_hack
 
-import i3ipc
 
-i3 = i3ipc.Connection()
-t = i3.get_tree()
-
-orig_win_id = t.find_focused().window
-
-# for tws
-win_names: list[str] = [
-    'Interactive Brokers',  # tws running in i3
-    'IB Gateway',  # gw running in i3
-    # 'IB',  # gw running in i3 (newer version?)
-]
-
-for name in win_names:
-    results = t.find_titled(name)
-    print(f'results for {name}: {results}')
-    if results:
-        con = results[0]
-        print(f'Resetting data feed for {name}')
-        win_id = str(con.window)
-        w, h = con.rect.width, con.rect.height
-
-        # TODO: seems to be a few libs for python but not sure
-        # if they support all the sub commands we need, order of
-        # most recent commit history:
-        # https://github.com/rr-/pyxdotool
-        # https://github.com/ShaneHutter/pyxdotool
-        # https://github.com/cphyc/pyxdotool
-
-        # TODO: only run the reconnect (2nd) kc on a detected
-        # disconnect?
-        for key_combo, timeout in [
-            # only required if we need a connection reset.
-            # ('ctrl+alt+r', 12),
-            # data feed reset.
-            ('ctrl+alt+f', 6)
-        ]:
-            subprocess.call([
-                'xdotool',
-                'windowactivate', '--sync', win_id,
-
-                # move mouse to bottom left of window (where there should
-                # be nothing to click).
-                'mousemove_relative', '--sync', str(w-4), str(h-4),
-
-                # NOTE: we may need to stick a `--retry 3` in here..
-                'click', '--window', win_id,
-                '--repeat', '3', '1',
-
-                # hackzorzes
-                'key', key_combo,
-                ],
-                timeout=timeout,
-            )
-
-# re-activate and focus original window
-subprocess.call([
-    'xdotool',
-    'windowactivate', '--sync', str(orig_win_id),
-    'click', '--window', str(orig_win_id), '1',
-])
+if __name__ == '__main__':
+    i3ipc_xdotool_manual_click_hack()
